Tangling already is handled by Doom
I didn't realize that doom has a literate module, which when enabled
auto-tangles config.org into config.el for us.

This means that we don't need the tangle header on the file, and we also
do not need the auto tangle on save method.

I've removed both and enabled the tangle module, including documentation
regarding it needing to be turned on.