Correct Olivetti setup, and disable org git gutter
This corrects how we setup Olivetti package. We were not setting it in a
way where the column width was being honored, which was my fault. I was
not establishing the config in an idiomatic way for doom Emacs.

Doom Emacs has several helpers to ensure the hooks and config settings
are established at the right point in the package life-cycle.

This ensures that:
1. Olivetti is run after org finishes being loaded
2. Turns on fancier margin styling
3. Adds the Olivetti mode hook in a way that fits within the config
life-cycle of packages within doom Emacs